[#3] - First steps towards sender implementation
Working on a sender which can automatically chunk the data into packets
and send it with its own sequencing logic.

Added a new method as well to calculate the size

However I'm rethinking the JSON-style approach and might try to use a
more rigid approach where we just define bytes in the first X locations
to be metadata like sequence/tag/etc.. then the rest is the actual data.

Still working out the kinks but with the tests going coming along with
development it shouldn't be too bad.