[#5] - Rewrite of consensus for non-MPI use
This is my first attempt at implementing consensus. I'm still trying to
translate some of the old code to try to get it to work. The corrective
part is currently what's throwing me off.

I may be able to implement normal consensus for now but as of this
commit it's not working.

I've tested a large majority of the new methods in this module. The
critical ones are all tested and seem to work. Integration tests (when
we get around to it) would be nice.

[Paper for corrective
consensus](https://www.researchgate.net/profile/Andreas_Terzis/publication/224220752_Corrective_Consensus_Converging_to_the_Exact_Average/links/0c96051d5e5faf1ff6000000.pdf)

Need to continue running tests and rewriting the run method and we
should be good to go....